优化 map 查询字段类型
